Fixes #5002: Package upgrade can fail if /etc/{cron.d,init.d,default}/ru...

+++ rudder-agent/SPECS/rudder-agent.spec
@@ -274,13 +274,13 @@ install -m 755 %{SOURCE7} %{buildroot}/opt/rudder/bin/check-rudder-agent
 # Do this only during upgrade process
 if [ $1 -eq 2 ];then
 	# Keep a backup copy of Rudder agent init and cron files to prevent http://www.rudder-project.org/redmine/issues/3995
-	mkdir -p /var/backups/rudder
-	cp -af /etc/init.d/rudder-agent /var/backups/rudder/rudder-agent.init-$(date +%Y%m%d)
-	echo "INFO: A back up copy of the /etc/init.d/rudder-agent has been created in /var/backups/rudder"
-	cp -af /etc/default/rudder-agent /var/backups/rudder/rudder-agent.default-$(date +%Y%m%d)
-	echo "INFO: A back up copy of the /etc/default/rudder-agent has been created in /var/backups/rudder"
-	cp -af /etc/cron.d/rudder-agent /var/backups/rudder/rudder-agent.cron-$(date +%Y%m%d)
-	echo "INFO: A back up copy of the /etc/cron.d/rudder-agent has been created in /var/backups/rudder"
+  for i in init.d default cron.d; do
+    if [ -f /etc/${i}/rudder-agent ]; then
+	    mkdir -p /var/backups/rudder
+	    cp -af /etc/${i}/rudder-agent /var/backups/rudder/rudder-agent.$(basename ${i} .d)-$(date +%Y%m%d)
+	    echo "INFO: A back up copy of /etc/${i}/rudder-agent has been created in /var/backups/rudder"
+    fi
+  done
 fi
 
 %post -n rudder-agent
@@ -433,14 +433,20 @@ fi
 #=================================================
 
 # Do it during upgrade and uninstall
+
 # Keep a backup copy of uuid.hive
-mkdir -p /var/backups/rudder
-cp -f /opt/rudder/etc/uuid.hive /var/backups/rudder/uuid-$(date +%Y%m%d).hive
-echo "INFO: A back up copy of the /opt/rudder/etc/uuid.hive has been created in /var/backups/rudder"
+if [ -f /opt/rudder/etc/uuid.hive ]; then
+  mkdir -p /var/backups/rudder
+  cp -f /opt/rudder/etc/uuid.hive /var/backups/rudder/uuid-$(date +%Y%m%d).hive
+  echo "INFO: A back up copy of the /opt/rudder/etc/uuid.hive has been created in /var/backups/rudder"
+fi
 
 # Keep a backup copy of CFEngine ppkeys
-cp -af /var/rudder/cfengine-community/ppkeys/ /var/backups/rudder/ppkeys-$(date +%Y%m%d)
-echo "INFO: A back up copy of the /var/rudder/cfengine-community/ppkeys has been created in /var/backups/rudder"
+if [ -d /var/rudder/cfengine-community/ppkeys/ ]; then
+  mkdir -p /var/backups/rudder
+  cp -af /var/rudder/cfengine-community/ppkeys/ /var/backups/rudder/ppkeys-$(date +%Y%m%d)
+  echo "INFO: A back up copy of the /var/rudder/cfengine-community/ppkeys has been created in /var/backups/rudder"
+fi
 
 
 %postun -n rudder-agent
